added check and censor functions
